Serial out memory bugfix
No more seg faults in serialize.cpp write function, and the data is accurately written to the file (as best as I can tell).  Still need to verify by testing an integration of serial in.

I just want to say, this was an incredibly hard issue to nail down and fix.  I broke GDB at one point.  So yeah.